feat: add pages to edit nodes and plugins for existing clusters
