+++ console-frontend/src/app/add-cluster-nodes/add-cluster-nodes.component.html
@@ -24,133 +24,12 @@
   </div>
 
   <!-- Form -->
-  <form [formGroup]="nodePoolsForm" (ngSubmit)="onSubmit()" class="space-y-6 px-6 py-6">
-    <!-- Node Pools -->
-    <div formArrayName="nodePools" class="space-y-6">
-      @for (nodePoolControl of nodePools.controls; track nodePoolControl) {
-        @let nameControl = nodePoolControl.get('name');
-        <div
-          [formGroupName]="$index"
-          class="rounded-lg border border-gray-200 bg-gray-50 p-4 dark:border-gray-800 dark:bg-gray-900"
-        >
-          <div class="mb-4 flex items-center justify-between">
-            <h3 class="text-lg font-medium dark:text-white">Node pool {{ $index + 1 }}</h3>
-            @if (nodePools.length > 1) {
-              <button
-                type="button"
-                (click)="removeNodePool($index)"
-                class="inline-flex cursor-pointer items-center rounded-md border border-rose-300 bg-white px-2 py-1 text-xs font-medium text-rose-700 shadow-sm ring-rose-500 hover:bg-rose-50 focus:ring-2 focus:ring-offset-2 focus:outline-none dark:border-rose-600 dark:bg-gray-900 dark:text-rose-300 dark:ring-offset-gray-950 dark:hover:bg-rose-900 dark:hover:text-white"
-              >
-                <svg xmlns="http://www.w3.org/2000/svg" class="mr-1 h-3 w-3" viewBox="0 0 24 24">
-                  <path
-                    fill="none"
-                    stroke="currentColor"
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    stroke-width="2"
-                    d="M6 18L18 6M6 6l12 12"
-                  />
-                </svg>
-                Remove
-              </button>
-            }
-          </div>
-
-          <div class="space-y-4">
-            <!-- Node Pool Name -->
-            <div>
-              <label
-                [for]="'nodePoolName-' + $index"
-                class="mb-2 block text-sm font-medium text-gray-700 dark:text-white"
-              >
-                Node pool name
-              </label>
-              <input
-                [id]="'nodePoolName-' + $index"
-                formControlName="name"
-                type="text"
-                #nodePoolNameInput
-                placeholder="E.g. my-pool-name"
-              />
-              @if (nameControl?.invalid && (nameControl?.dirty || nameControl?.touched)) {
-                <p class="mt-2 text-sm text-rose-600 dark:text-rose-400">
-                  {{ getNodePoolNameError($index) }}
-                </p>
-              }
-            </div>
-
-            <!-- Machine Type -->
-            <div>
-              <label
-                [for]="'machineType-' + $index"
-                class="mb-2 block text-sm font-medium text-gray-700 dark:text-white"
-              >
-                Machine type
-              </label>
-              <select [id]="'machineType-' + $index" formControlName="machineType">
-                @for (mt of machineTypes; track mt.value) {
-                  <option [value]="mt.value">{{ mt.label }}</option>
-                }
-              </select>
-            </div>
-
-            <!-- Autoscaling -->
-            <div class="grid grid-cols-1 gap-4 sm:grid-cols-2">
-              <div>
-                <label
-                  [for]="'autoscaleMin-' + $index"
-                  class="mb-2 block text-sm font-medium text-gray-700 dark:text-white"
-                >
-                  Autoscale minimum
-                </label>
-                <input
-                  [id]="'autoscaleMin-' + $index"
-                  formControlName="autoscaleMin"
-                  type="number"
-                  min="1"
-                  max="100"
-                />
-              </div>
-
-              <div>
-                <label
-                  [for]="'autoscaleMax-' + $index"
-                  class="mb-2 block text-sm font-medium text-gray-700 dark:text-white"
-                >
-                  Autoscale maximum
-                </label>
-                <input
-                  [id]="'autoscaleMax-' + $index"
-                  formControlName="autoscaleMax"
-                  type="number"
-                  min="1"
-                  max="100"
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-      }
-    </div>
-
-    <button type="button" (click)="addNodePool()" class="btn-light inline-flex items-center">
-      <svg xmlns="http://www.w3.org/2000/svg" class="mr-1.5 h-4 w-4" viewBox="0 0 24 24">
-        <path
-          fill="none"
-          stroke="currentColor"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-          stroke-width="2"
-          d="M12 5v14m-7-7h14"
-        />
-      </svg>
-      Add another node pool
-    </button>
-
-    <!-- Form Actions -->
-    <div class="flex space-x-3 border-t border-gray-200 pt-6 dark:border-gray-800">
-      <a routerLink="/add-cluster" class="btn-secondary"> Previous step </a>
-      <button type="submit" class="btn-primary">Next step</button>
-    </div>
-  </form>
+  <div class="px-6 py-6">
+    <app-shared-node-pools-form (formSubmit)="onFormSubmit($event)">
+      <div class="flex space-x-3">
+        <a routerLink="/add-cluster" class="btn-secondary"> Previous step </a>
+        <button type="submit" class="btn-primary">Next step</button>
+      </div>
+    </app-shared-node-pools-form>
+  </div>
 </div>
