+++ console-frontend/src/app/cluster-overview/cluster-overview.component.html
@@ -298,7 +298,7 @@
         <h3 class="mb-4 text-lg font-medium dark:text-white">
           Members
           <a
-            routerLink="/cluster-members"
+            routerLink="/project-members"
             class="btn-light ml-1.5 inline-flex items-center rounded! px-2! py-1! text-xs! shadow-xs!"
           >
             <svg xmlns="http://www.w3.org/2000/svg" class="mr-1 h-4 w-4" viewBox="0 0 24 24">
@@ -357,7 +357,7 @@
         <h3 class="mb-4 text-lg font-medium dark:text-white">
           Node pools
           <a
-            routerLink="/add-cluster-nodes"
+            routerLink="/cluster-nodes"
             class="btn-light ml-1.5 inline-flex items-center rounded! px-2! py-1! text-xs! shadow-xs!"
           >
             <svg xmlns="http://www.w3.org/2000/svg" class="mr-1 h-4 w-4" viewBox="0 0 24 24">
@@ -416,7 +416,7 @@
         <h3 class="mb-4 text-lg font-medium dark:text-white">
           Plugins
           <a
-            routerLink="/add-cluster-plugins"
+            routerLink="/cluster-plugins"
             class="btn-light ml-1.5 inline-flex items-center rounded! px-2! py-1! text-xs! shadow-xs!"
           >
             <svg xmlns="http://www.w3.org/2000/svg" class="mr-1 h-4 w-4" viewBox="0 0 24 24">
