+++ console-frontend/src/app/cluster-plugins/cluster-plugins.component.css
@@ -1,0 +1,1 @@
+/* Component-specific styles can be added here if needed */
