+++ console-frontend/src/app/cluster-plugins/cluster-plugins.component.html
@@ -1,0 +1,19 @@
+<div class="rounded-md border border-gray-200 bg-white dark:border-gray-800 dark:bg-gray-950">
+  <!-- Header -->
+  <div class="border-b border-gray-200 px-6 py-4 dark:border-gray-800">
+    <h1 class="text-2xl font-bold dark:text-white">Cluster plugins</h1>
+    <p class="mt-1 text-sm text-gray-600 dark:text-gray-300">
+      Configure plugins for cluster 'cluster-1'.
+    </p>
+  </div>
+
+  <!-- Form -->
+  <div class="px-6 py-6">
+    <app-shared-plugins-form (formSubmit)="onFormSubmit($event)">
+      <div class="flex space-x-3">
+        <button type="button" (click)="onCancel()" class="btn-secondary">Cancel</button>
+        <button type="submit" class="btn-primary">Save changes</button>
+      </div>
+    </app-shared-plugins-form>
+  </div>
+</div>
