+++ console-frontend/src/app/shared-node-pools-form/shared-node-pools-form.component.css
@@ -1,0 +1,1 @@
+/* Component-specific styles can be added here if needed */
