+++ console-frontend/src/app/shared-plugins-form/shared-plugins-form.component.css
@@ -1,0 +1,1 @@
+/* Component-specific styles can be added here if needed */
